[Refactor]: 리뷰코드 컴포넌트화해서 map으로 사용할수있게 함
